services/reserva/obtener-reserva/requirements.txt
sqlalchemy==1.4.*
cloud-sql-python-connector==0.9.*
sqlalchemy_pytds==0.3.5
python-tds==1.15.0
pytz==2024.1
pyjwt==2.8.0